backend/requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.36
PyMySQL==1.1.0
python-dotenv==1.0.1
cryptography==42.0.5
Flask-JWT-Extended==4.7.1
flask_cors==6.0.2